requirements.txt
langchain 
langchain_groq 
fastapi 
pydantic 
streamlit 
python-dotenv
uvicorn
langchain-community
psycopg2 
python-multipart